requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
motor>=3.1.1
python-dotenv>=1.0.0
pydantic>=1.10.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0